refactor(pubsub): move class to its own file
Nothing to see here folks, move along.
