+++ google/cloud/pubsub/pubsub_client.bzl
@@ -26,6 +26,7 @@ pubsub_client_hdrs = [
     "internal/publisher_logging.h",
     "internal/publisher_metadata.h",
     "internal/publisher_stub.h",
+    "internal/session_shutdown_manager.h",
     "internal/subscriber_logging.h",
     "internal/subscriber_metadata.h",
     "internal/subscriber_stub.h",
@@ -63,6 +64,7 @@ pubsub_client_srcs = [
     "internal/publisher_logging.cc",
     "internal/publisher_metadata.cc",
     "internal/publisher_stub.cc",
+    "internal/session_shutdown_manager.cc",
     "internal/subscriber_logging.cc",
     "internal/subscriber_metadata.cc",
     "internal/subscriber_stub.cc",
