+++ google/cloud/pubsub/pubsub_client_unit_tests.bzl
@@ -24,6 +24,7 @@ pubsub_client_unit_tests = [
     "internal/ordering_key_publisher_connection_test.cc",
     "internal/publisher_logging_test.cc",
     "internal/publisher_metadata_test.cc",
+    "internal/session_shutdown_manager_test.cc",
     "internal/subscriber_logging_test.cc",
     "internal/subscriber_metadata_test.cc",
     "internal/subscription_concurrency_control_test.cc",
